Update data-validation.md
I don't know if the above example should stay, at least this led to confusion as I couldn't get it to work in either V3.7.1 or 4.0.0. 

Therefore, the section "Configuration via Web UI" describes how it works and is also comprehensible for a non-expert. 

Changes are based on the experiences of this discussion.
https://github.com/netbox-community/netbox/discussions/16021